Phase 1 Step 2: extract SentimentService, kill duplicate FinBERT loader
- quantcore/services/sentiment.py: NewsCollector moved wholesale from
  fastMCPTest/news_collector.py (RSS fetch, yfinance news via gateway,
  FinBERT AutoModel loader with full prob breakdown) plus SentimentService:
  collect_news / get_news_sentiment / get_sentiment_trend /
  list_news_symbols (from news_sentiment_server), get_news (from
  stock_price_server), get_security_news + get_sentiment_dashboard (from
  the two REST news routes).
- FinBERT duplicate eliminated: stock_price_server's transformers-pipeline
  loader deleted; the single AutoModel loader produces byte-identical
  sentiment scores (verified in parity diff).
- YFinanceGateway gains news(); registry wires SentimentService.
- news_sentiment_server.py now a thin adapter (docstrings verbatim);
  news_collector.py is a re-export shim until Step 8 (options_analysis).
- api/app.py: /api/securities/<t>/news and /news/sentiment-summary are one
  service call deep; snapshot persistence moved inside the service.

Parity vs pre-refactor baseline (AAPL, all 7 surfaces incl. REST via test
client): identical except snapshot captured_at, which advances on every
call by design. 30 unit tests green against the test DB; 5 MCP servers
import; news server registers its 4 tools; Flask boots 38 routes.
