+++ docs/proposals/phase1-migration-plan.md
@@ -10,7 +10,7 @@ Update this table as part of every step-commit. Any dev machine can resume by re
 | -1 | Safety setup (prod backup, tag, test DB seed, DSN guard, flyway) | DONE | — | 2026-06-12 | Prod backup backups/quantcore_prod_2026-06-12.dump (5.6 MB) restore-verified into local scratch DB quantcore_restore_check AND used to reseed the test DB (exact prod parity: 16 tables, 176,998 options_contracts, 114,221 ohlcv). Test Cloud SQL instance quantcore-test-20260606 started; proxy on 5434. Flyway 12.8.1 installed; db/flyway.conf + db/migrations/ created; flyway info connects (baseline V1 on first migrate). db_safety guard hardened (also catches quantcore.db imported pre-override) and wired into test_options_contract_tools.py — NOTE: bare `python -m unittest discover` previously hit PROD via default DSN; guard now redirects/refuses. Worktree skipped per user — working in main checkout |
 | 0 | Scaffolding (packages, repositories move, registry) | DONE | — | 2026-06-12 | quantcore/{gateways,repositories,analytics,services} created; 6 stores moved to repositories (OhlcvRepository/FundamentalsRepository OO facades added); re-export shims at old fastMCPTest paths; services/registry.py composition root (lazy lru_cache); notifier.py + test_options_contract_tools.py import from quantcore directly. Verified: 30 tests green vs test DB, 5 MCP servers import, Flask 38 routes, registry wires. Note: the file renames were accidentally staged into the Step -1 commit (613cdee) — that single commit is transiently broken; this commit restores coherence |
 | 1 | Microstructure service | DONE | — | 2026-06-12 | Template step. quantcore/gateways/yfinance_gateway.py (YFinanceGateway: ticker_info w/ 15s timeout, fast_info, expirations, option_chain) + quantcore/services/microstructure.py (MicrostructureService: get_short_interest/get_dark_pool/get_bid_ask_spread, bodies verbatim); registry wires gateway+service; market_analysis_server.py now a thin adapter (3 one-line tools, docstrings verbatim); stock_price_server.py cross-import repointed at the service (lines ~3179/3195/3212). fastmcp 3.2.3 confirmed: @mcp.tool() returns the original fn, so the old cross-calls were plain delegation — repoint is exact. Parity vs /tmp/micro_parity_before.json (AAPL): all fields identical incl. live-quote ones. ValueError contract preserved. 30 tests green vs test DB; 5 MCP servers import; Flask 38 routes |
-| 2 | Sentiment/News service | PENDING | — | — | |
+| 2 | Sentiment/News service | DONE | — | 2026-06-12 | quantcore/services/sentiment.py: NewsCollector moved wholesale from fastMCPTest/news_collector.py (RSS + yfinance fetch, FinBERT AutoModel loader) + SentimentService (collect_news/get_news_sentiment/get_sentiment_trend/list_news_symbols from news_sentiment_server; get_news from stock_price_server; get_security_news + get_sentiment_dashboard from REST routes). FinBERT duplicate KILLED: stock_price_server pipeline loader deleted; single loader in the service (scores byte-identical in parity check). YFinanceGateway.news() added. news_sentiment_server.py thin adapter (4 one-line tools); news_collector.py is a shim (options_analysis imports NewsCollector until Step 8); api/app.py /news + /news/sentiment-summary one call deep; stock_price_server internal news call repointed. Parity vs /tmp/sentiment_parity_before.json: only drift = snapshot captured_at (advances per call by design). 30 tests green; 5 servers import; Flask 38 routes |
 | 3 | Fundamentals service | PENDING | — | — | |
 | 4 | Prices/technicals service + analytics | PENDING | — | — | |
 | 5 | Options service + PolygonGateway | PENDING | — | — | |
